Renamed files and variables to plural. Removed a few arguments on deliveries delete function jpedrodsp#25
